requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
addict=2.4.0=pypi_0
altair=5.4.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=24.2.0=pypi_0
bleach=6.2.0=pypi_0
blinker=1.8.2=pypi_0
bokeh=3.6.1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
configargparse=1.7=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
dash=2.18.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.8.7=py311hfdbb021_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
fastjsonschema=2.20.0=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.10.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.29.0=pyh707e725_0
ipywidgets=8.1.5=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h143b758_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
linkify-it-py=2.0.3=pypi_0
lxml=5.3.0=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdit-py-plugins=0.4.2=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
narwhals=1.13.3=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pypi_0
nibabel=5.3.2=pypi_0
nilearn=0.10.4=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
open3d=0.18.0=pypi_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=pypi_0
panel=1.5.3=pypi_0
param=2.1.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=11.0.0=pypi_0
pip=24.2=py311h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_0
plotly=5.24.1=pypi_0
pooch=1.8.2=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
protobuf=5.28.3=pypi_0
psutil=6.1.0=py311h9ecbd09_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=18.0.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.2.0=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.11.10=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.11=2_cp311
pytz=2024.2=pypi_0
pyvista=0.44.1=pypi_0
pyviz-comms=3.0.3=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py311h7deb3e3_3
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
retrying=1.3.4=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.20.1=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
scooby=0.10.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py311h06a4308_0
simpleitk=2.4.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.4=pypi_0
stpyvista=0.1.4=pypi_0
streamlit=1.40.0=pypi_0
streamlit-option-menu=0.4.0=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
tk=8.6.14=h39e8969_0
toml=0.10.2=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.1=py311h9ecbd09_1
tqdm=4.66.6=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
trimesh=4.5.1=pypi_0
triton=3.1.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
uc-micro-py=1.0.3=pypi_0
urllib3=2.2.3=pypi_0
vtk=9.3.1=pypi_0
watchdog=5.0.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py311h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xyzservices=2024.9.0=pypi_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h3b0a872_6
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
